// src/components/training/anatomical/types.ts
export type Slug =
  | "abs"
  | "adductors"
  | "ankles"
  | "back-top"
  | "biceps"
  | "calves"
  | "chest"
  | "deltoids"
  | "feet"
  | "forearm"
  | "gluteal"
  | "gluteus"
  | "hamstring"
  | "hands"
  | "hair"
  | "head"
  | "knees"
  | "lower-back"
  | "neck"
  | "obliques"
  | "cuadriceps"
  | "tibialis"
  | "trapezius"
  | "triceps"
  | "upper-back"
  | "pecho"
  | "oblicuos"
  | "abdominales"
  | "cuello"
  | "trapecio"
  | "deltoides"
  | "aductores"
  | "rodillas"
  | "antebrazos"
  | "manos"
  | "tobillos"
  | "pies"
  | "cabeza"
  | "pelo"
  | "espalda-superior"
  | "espalda-inferior"
  | "gluteos"
  | "isquiotibiales"
  | "pantorrillas"
  | "mano"
  | "pie"
  | "tibial"
  | "tobillo";

export interface BodyPart {
  slug: Slug;
  color?: string;
  path: {
    common?: string[];
    left?: string[];
    right?: string[];
  };
}